projects/bookshelf/pyproject.toml
[project]
name = "bookshelf"
version = "0.1.0"
description = "Multi-agent AI assistant that turns a Malaysian book shop owner's sales data into product decisions."
requires-python = ">=3.11,<3.14"
dependencies = [
    "google-adk>=1.27.4,<2.0.0",
    "a2a-sdk==0.3.*",
    "google-genai>=0.5.0",
    "google-auth>=2.30.0",
    "fastapi==0.124.*",
    "uvicorn==0.40.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "pydantic>=2.0.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "google-cloud-logging==3.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]
